requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5

# Data Processing
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3

# API Requests
requests==2.31.0
urllib3==2.0.7

# Geospatial
geopandas==0.13.2
shapely==2.0.2
rasterio==1.3.8
folium==0.14.0

# Weather & Climate APIs
pyowm==3.3.0
python-dateutil==2.8.2

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# Notifications
twilio==8.5.0
pyttsx3==2.90

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.2
black==23.7.0
flake8==6.0.0
